requirements.txt
absl-py==2.3.1
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
black==21.4b2
certifi==2025.8.3
charset-normalizer==3.4.3
cityscapesScripts==2.2.2
click==8.1.8
cloudpickle==3.1.1
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
coverage==7.10.7
cycler==0.12.1
Cython==3.0.2
decorator==5.2.1
deepspeed==0.10.3
detectron2 @ git+https://github.com/MaureenZOU/detectron2-xyz.git@42121d75e10d9f858f3a91b6a39f5722c02868f0
diffdist==0.1
docker-pycreds==0.4.0
einops @ git+https://github.com/arogozhnikov/einops.git@361b11e87da94ead4bd09de636c5dbed73e0e3e0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
flake8==7.3.0
fonttools==4.59.1
fsspec==2025.7.0
ftfy==6.1.1
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.45
gradio==3.42.0
gradio_client==0.5.0
grpcio==1.74.0
h11==0.16.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
infinibatch==0.1.1
iniconfig==2.1.0
interrogate==1.7.0
iopath==0.1.9
ipython==8.18.1
isort==4.3.21
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json-tricks==3.17.3
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
kornia==0.7.0
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.7.3
loguru==0.7.3
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.2.1
mccabe==0.7.0
mmcv-full @ https://download.openmmlab.com/mmcv/dist/cu121/torch2.1.0/mmcv_full-1.7.2-cp39-cp39-manylinux1_x86_64.whl#sha256=2d3502534b10a72892a65f25345b159942687ef89b01fd2ecdc6ddeb806b21e6
more-itertools==10.7.0
mpi4py==4.1.0
mpmath==1.3.0
mup==1.0.0
mypy_extensions==1.1.0
narwhals==2.1.2
networkx==3.2.1
ninja==1.13.0
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==3.1.0
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencv-python==4.8.1.78
opencv-python-headless==4.12.0.88
orjson==3.11.2
packaging==25.0
pandas==2.0.3
parso==0.8.5
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.2.0
prettytable==3.16.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==14.0.2
pybind11==3.0.1
pycocotools==2.0.7
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
pyquaternion==0.9.9
pytest==8.4.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-metric-learning==2.9.0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.1
referencing==0.36.2
regex==2023.10.3
requests==2.32.5
rpds-py==0.27.0
safetensors==0.6.2
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.35.0
setproctitle==1.3.6
Shapely==1.8.0
simsimd==6.5.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
stringzilla==3.12.6
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.11.0
timm==1.0.21
tokenizers==0.14.1
toml==0.10.2
tomli==2.3.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.34.0
triton==2.1.0
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vision-datasets==0.2.2
wandb==0.21.1
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
xdoctest==1.3.0
yacs==0.1.8
yapf==0.43.0
zipp==3.23.0
